Creating debian release process

+++ scripts/release-debianpackage.sh
@@ -25,10 +25,11 @@ echo "=> Creating apt repo folder"
 mkdir ./aptcvecli
 echo "=> Syncing S3 bucket locally"
 aws s3 sync s3://aptthepublicclouds/cvecli ./aptcvecli --debug
-echo "=> Creating pools directory"
+echo "=> Checking pools directory"
 if [ -d "./aptcvecli/pool/main" ]; then
   echo "=> ./aptcvecli/pool/main already exists"
 else
+  echo "=> Creating pools directory"
   mkdir -p ./aptcvecli/pool/main
 fi
 for architecture in "${architectures[@]}"; do
@@ -36,7 +37,7 @@ for architecture in "${architectures[@]}"; do
   releaseArchitectures+=" "
   echo "=> Moving $architecture debian package to local apt repo"
   mv "./cvecli_$version-1_$architecture.deb" "./aptcvecli/pool/main/cvecli_$version-1_$architecture.deb"
-  echo "=> Creating $architecture packages directory"
+  echo "=> Checking for $architecture packages directory"
   if [ -d "./aptcvecli/dists/stable/main/binary-$architecture" ]; then
     echo "=> ./aptcvecli/dists/stable/main/binary-$architecture already exists"
   else
@@ -60,10 +61,25 @@ for architecture in "${architectures[@]}"; do
   echo "=> Compressing $architecture package file"
   gzip -k "./aptcvecli/dists/stable/main/binary-$architecture/Packages"
 done
-echo "=> Removing old release files"
-rm ./aptcvecli/dists/stable/Release
-rm ./aptcvecli/dists/stable/Release.gpg
-rm ./aptcvecli/dists/stable/InRelease
+echo "=> Checking for old release files"
+if [ -f "./aptcvecli/dists/stable/Release" ]; then
+  echo "=> Removing ./aptcvecli/dists/stable/Release"
+  rm ./aptcvecli/dists/stable/Release
+else
+  echo "=> ./aptcvecli/dists/stable/Release does not exist"
+fi
+if [ -f "./aptcvecli/dists/stable/Release.gpg" ]; then
+  echo "=> Removing ./aptcvecli/dists/stable/Release.gpg"
+  rm ./aptcvecli/dists/stable/Release.gpg
+else
+  echo "=> ./aptcvecli/dists/stable/Release.gpg does not exist"
+fi
+if [ -f "./aptcvecli/dists/stable/InRelease" ]; then
+  echo "=> Removing ./aptcvecli/dists/stable/InRelease"
+  rm ./aptcvecli/dists/stable/InRelease
+else
+  echo "=> ./aptcvecli/dists/stable/InRelease does not exist"
+fi
 ./aptcvecli/dists/stable/Release << EOF
 Origin: apt.thepublicclouds.com
 Suite: stable
